refactor: frontend を colocation 構成へ + features/village を villages に改名 (step-4.1)
- routes/ は画面軸 (colocation): 画面専用 UI/logic はルートにコロケート。
  home を routes/home/{route.tsx, MenuTile.tsx} に再編 (features/home は廃止)
- features/ は機能(ドメイン)軸 (複数画面で共有): 一覧ドメインを features/village →
  features/villages (複数形) に改名。単数 village は今後の村詳細ドメイン用に確保
- frontend/README.md にディレクトリ構成/設計方針を明記
